==> 1. Introduction & Data Cleanup.ipynb <==
# FLIGHT DELAY VISUALIZATION & MODELING

## By Ravi Surdhar, Dec 2018

This notebook uses the 2015 Flight Delays Dataset from Kaggle, as well as the Iowa State University AWOS Download Service, and attempts to generate a model to predict flight delays and cancellations based on airport weather. 

The tech stack includes Pandas, Seaborn, Matplotlib, Scikit-Learn, and TensorFlow. Topics explored include: identifying and fixing corrupted data, big data techniques and optimizations, web scraping, data visualization, and model selection and evaluation.

### OVERVIEW
#### 1. Introduction & Data Cleanup
*Fixing Corrupted Data*

*Reduction & Optimization*
#### 2. Web Scraping & Adding Weather Data
#### 3. Visualization
#### 4. Modeling

## Data Cleanup

First, let's import the usual suspects.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')
%matplotlib inline
pd.set_option('display.max_columns', 35)

flights = pd.read_csv('flights.csv')
airport_iata_id = pd.read_csv('airport_iata_id.csv', index_col=0)
print(flights.info(memory_usage='deep'))
flights.describe()

flights.head()

Wow, 2.6 GB! Let's start by whittling that down. First, we can drop columns that we don't think will provide any correlation to how delayed a flight will be.

flights_reduced1 = flights.drop(['FLIGHT_NUMBER', 'TAIL_NUMBER', 'TAXI_OUT', 'WHEELS_OFF', 'SCHEDULED_TIME', 'ELAPSED_TIME', 'AIR_TIME', 'DISTANCE', 'WHEELS_ON', 'TAXI_IN', 'AIR_SYSTEM_DELAY', 'SECURITY_DELAY', 'AIRLINE_DELAY', 'LATE_AIRCRAFT_DELAY', 'WEATHER_DELAY'], axis=1)
flights_reduced1.info(memory_usage='deep')

That saved 1 GB right there! Good start! We'll continue to do some more memory optimizations in a moment, but for now, let's make sure our dataframe has no errors by making sure all the numeric columns make sense, and the string columns all contain what we expect. 

print(flights_reduced1.describe())
print(flights_reduced1['AIRLINE'].unique())
print(flights_reduced1['ORIGIN_AIRPORT'].unique())
print(flights_reduced1['DESTINATION_AIRPORT'].unique())
print(flights_reduced1['CANCELLATION_REASON'].unique())

The numeric columns all make sense: all of the years are 2015, the months go from 1 to 12 with a mean of 7, etc. There's also the same number of entries in the YEAR column as there are in the MONTH, DAY, DAY_OF_WEEK, SCHEDULED_DEPARTURE, SCHEDULED_ARRIVAL, and CANCELLED columns, just as we'd expect (some flights wouldn't have a DEPARTURE_TIME, DEPARTURE_DELAY, etc if they're cancelled). CANCELLED makes sense too; it's a one-hot encoded, with 1 indicating that the flight was cancelled.

For the string columns, CANCELLATION_REASON is what we expect, where the different letters correspond to the reason for a cancellation, and a NaN value if the flight was not cancelled. AIRLINE is also what we'd expect, as those are two-letter identifiers that are given in the airlines.csv file. 

However, the airports are clearly corrupted. All of the entries in ORIGIN_AIRPORT and DESTINATION_AIRPORT should be three-letter IATA airport identifiers, but there are a bunch of strings of five numbers, as well as a bunch of five-digit integers. Let's fix that next.

After quite a bit of digging, it turns out that the corrupted airport entries are only for flights in Oct. So let's start by isolating those.

corrupted_flights = flights_reduced1[flights_reduced1['MONTH'] ==10]
corrupted_flights.head()

Luckily, the person from the Department of Transportation who uploaded the dataset included an Airport ID.csv file, which contains a column of five-digit airport IDs and a column of names. We can merge this with a list of all airports with the IATA codes (also from the DOT) to create a dictionary of key : value pairs corresponding to ID : IATA.

airport_id = pd.read_csv('AIRPORT ID.csv')
airport_iata = pd.read_csv('AIRPORT IATA.csv')
airport_iata_id = pd.merge(airport_iata, airport_id, on='Description')
airport_iata_id.rename({'Code_x': 'IATA', 'Code_y': 'ID'}, axis=1, inplace=True)
airport_dictionary = airport_iata_id[['ID', 'IATA']].set_index('ID').to_dict()['IATA']

Before we can use the dictionary to replace the corrupted airports, we first need to convert all of airports that are five-digit strings to five-digit integers, because our dictionary keys are all integers. Then we can run the replacement. With tens of thousands of flights in our corrupted_flights dataframe, this replacement takes a few minutes!

corrupted_flights['ORIGIN_AIRPORT'] = pd.to_numeric(corrupted_flights['ORIGIN_AIRPORT'])
corrupted_flights['DESTINATION_AIRPORT'] = pd.to_numeric(corrupted_flights['DESTINATION_AIRPORT'])
fixed_flights = corrupted_flights.replace({'ORIGIN_AIRPORT' : airport_dictionary, 'DESTINATION_AIRPORT' : airport_dictionary})

Finally, we can merge our fixed_flights with our flights_reduced1 dataframe, then do a couple checks on the results.

merged_flights = flights_reduced1[flights_reduced1['MONTH'] < 10].append(fixed_flights).append(flights_reduced1[flights_reduced1['MONTH'] > 10])
print(merged_flights['MONTH'].unique(), '\n',               # Expect to see all 12 months
    merged_flights['ORIGIN_AIRPORT'].unique(), '\n',        # Expect to see only 3-letter IATA strings
    merged_flights['ORIGIN_AIRPORT'].nunique(), '\n',
    merged_flights['DESTINATION_AIRPORT'].unique(), '\n',   # Expect to see only 3-letter IATA strings
    merged_flights['DESTINATION_AIRPORT'].nunique())

Looks good! Our next step will be to only include flights in our dataframe that are the in the top 1000 most travelled routes. Airports with only a few flights will most likely create outliers in our data that will throw off our modeling. 

flights_grouped = flights.groupby(['ORIGIN_AIRPORT', 'DESTINATION_AIRPORT'])
top_1000 = flights_grouped.size().reset_index(name='counts').sort_values(by='counts', ascending=False).head(1000)
top_airports = np.unique(np.append(top_1000['ORIGIN_AIRPORT'].unique(), top_1000['DESTINATION_AIRPORT'].unique()))
flights_reduced2 = flights_reduced1[flights_reduced1['ORIGIN_AIRPORT'].isin(top_airports) & flights_reduced1['DESTINATION_AIRPORT'].isin(top_airports)]
print(top_1000.head())
print(top_1000.tail())

This looks reasonable. Our least common route in our dataframe has fewer than 5 flights per day, and our most common has 37 flights per day, so this spans a little less than one order of magnitude. (In retrospect, using only the top 1000 routes is a little arbitrary. I could have used only routes within one order of magnitude of the most common route. Also, there are 27132 routes with less than one flight per day!) Let's take a look at our new flights_reduced2 dataframe.

print(flights_reduced2.describe())
print(flights_reduced2.info(memory_usage='deep'))

We've reduced the number of rows by nearly 20% and the memory by 13%, but we can still do a lot better!


## Reduction & Optimization

One big consideration when working with large dataframes is to be efficient with your data types. Here, we can see that YEAR, MONTH, DAY, and DAY_OF_WEEK are all 8-byte int64, which is way bigger than we need. Even our one-hot CANCELLED column is an int64! Let's start by downcasting all of the numeric rows (note that the times are all integer values,  but they're required to be floats as ints cannot store NaN values.

flights_reduced2_int = flights_reduced2.select_dtypes(include=['int64'])
converted_int = flights_reduced2_int.apply(pd.to_numeric,downcast='unsigned')

flights_reduced2_float = flights_reduced2.select_dtypes(include=['float64'])
converted_float = flights_reduced2_float.apply(pd.to_numeric,downcast='float')

optimized_flights_reduced2 = flights_reduced2.copy()

optimized_flights_reduced2[converted_int.columns] = converted_int
optimized_flights_reduced2[converted_float.columns] = converted_float

print(flights_reduced2.info(memory_usage='deep', verbose=False))
print(optimized_flights_reduced2.info(memory_usage='deep', verbose=False))

That's another 22% reduction! Not bad, but we can do better. Way better. 

Our four Object (string) columns have very few unique entries in each column. That's a perfect canditate to converting to a Catagory type, which Pandas treats as an int under the hood (see this great article for more info: https://www.dataquest.io/blog/pandas-big-data/)

flights_reduced2_obj = flights_reduced2.select_dtypes(include=['object']).copy()
print(flights_reduced2_obj.describe())

flights_cat = flights_reduced2_obj.astype('category')
print(flights_reduced2_obj.info(memory_usage='deep', verbose=False))
print(flights_cat.info(memory_usage='deep', verbose=False))

997 MB down to 64 MB! That's a 94% reduction! But we still have one more trick up our sleeve!

We can combine the first 5 columns of our dataframe (YEAR, MONTH, DAY, DAY_OF_WEEK, and SCHEDULED_DEPARTURE) into a single Pandas Timestamp object (Panda's implementation of Python's DateTime type). We can do the same for the SCHEDULED_ARRIVAL. Then we can drop the DEPARTURE_TIME and ARRIVAL_TIME columns, since those columns are equal to SCHEDULED_DEPARTURE + DEPARTURE_DELAY and SCHEDULED_ARRIVAL + ARRIVAL_DELAY, respectively.

# Make a df containing the Year, Month, Day, Hour, and Minute of the Scheduled Departure time, then convert that to a single column of Timestamps
df = pd.concat([flights_reduced2[['YEAR', 'MONTH', 'DAY']], np.floor(flights_reduced2['SCHEDULED_DEPARTURE']/100), flights_reduced2['SCHEDULED_DEPARTURE']%100], axis=1)
df.columns = ['Year', 'Month', 'Day', 'Hour', 'Minute']
flights_reduced2['SCHEDULED_DEPARTURE_DATETIME'] = pd.to_datetime(df)

# Do the same for the Scheduled Arrival time
df = pd.concat([flights_reduced2[['YEAR', 'MONTH', 'DAY']], np.floor(flights_reduced2['SCHEDULED_ARRIVAL']/100), flights_reduced2['SCHEDULED_ARRIVAL']%100], axis=1)
df.columns = ['Year', 'Month', 'Day', 'Hour', 'Minute']
flights_reduced2['SCHEDULED_ARRIVAL_DATETIME'] = pd.to_datetime(df)

flights_reduced2.drop(['YEAR', 'MONTH', 'DAY', 'DAY_OF_WEEK', 'SCHEDULED_DEPARTURE', 'DEPARTURE_TIME', 'SCHEDULED_ARRIVAL', 'ARRIVAL_TIME'], axis=1, inplace=True)
flights_reduced2.rename({'SCHEDULED_DEPARTURE_DATETIME' : 'SCHEDULED_DEPARTURE', 'SCHEDULED_ARRIVAL_DATETIME' : 'SCHEDULED_ARRIVAL'}, axis=1, inplace=True)
flights_reduced2.head()

At this point, it would be convenient to write a new .csv file with only our flights_reduced2 columns. We can also store all of the optimized data types into a dictionary so that if we have to shutdown the python kernen and start it again, we can read flights.csv directly into the optimized data types. 

flights_reduced2.to_csv('flights_reduced.csv')
optimized_flights_reduced2.drop(['YEAR', 'MONTH', 'DAY', 'DAY_OF_WEEK', 'SCHEDULED_DEPARTURE', 'DEPARTURE_TIME', 'SCHEDULED_ARRIVAL', 'ARRIVAL_TIME'], axis=1, inplace=True)
dtypes = optimized_flights_reduced2.dtypes
dtypes_col = dtypes.index
dtypes_type = [i.name for i in dtypes.values]
column_types = dict(zip(dtypes_col, dtypes_type))
print(dtypes_col , '\n', column_types)

Let's read that .csv back in as a new flights dataframe.

flights = pd.read_csv('flights_reduced.csv', 
                      index_col=0, 
                      dtype = {'AIRLINE': 'category', 
                                'ORIGIN_AIRPORT': 'category', 
                                'DESTINATION_AIRPORT': 'category', 
                                'DEPARTURE_DELAY': 'float32', 
                                'ARRIVAL_DELAY': 'float32', 
                                'DIVERTED': 'uint8', 
                                'CANCELLED': 'uint8', 
                                'CANCELLATION_REASON': 'category'}, 
                      parse_dates = ['SCHEDULED_DEPARTURE', 'SCHEDULED_ARRIVAL']).reset_index(drop=True)

flights.info(memory_usage='deep')

181.8 MB down from our original 2.6 GB! Our optimized flights dataframe uses 7% of the memory that our original one did!

In the next notebook, we'll add weather data for each flight into our dataframe.
